Refactor instructions into manageable files
Previously everything was in one file to do janky stuff
with numering opcodes. Now we have a proper handler for
this so we can split them into separate files.

TODO: We still need to update the instruction_modules list in
the instructions/__init__.py file
every time we add a new file.